agents/requirements.txt
flask
crewai
requests
langchain
groq
langchain-groq
python-dotenv
litellm
